Cleanup TMC Stepper a bit.
