+++ Marlin/stepper.cpp
@@ -87,37 +87,7 @@ static bool check_endstops = true;
 volatile long count_position[NUM_AXIS] = { 0, 0, 0, 0};
 volatile signed char count_direction[NUM_AXIS] = { 1, 1, 1, 1};
 
-// Stepper objects of TMC steppers are used
-#ifdef X_IS_TMC
-	TMC26XStepper stepperX(200,X_ENABLE_PIN,X_STEP_PIN,X_DIR_PIN,X_MAX_CURRENT,X_SENSE_RESISTOR);
-#endif
-#ifdef X2_IS_TMC
-	TMC26XStepper stepperX2(200,X2_ENABLE_PIN,X2_STEP_PIN,X2_DIR_PIN,X2_MAX_CURRENT,X2_SENSE_RESISTOR);
-#endif
-#ifdef Y_IS_TMC
-	TMC26XStepper stepperY(200,Y_ENABLE_PIN,Y_STEP_PIN,Y_DIR_PIN,Y_MAX_CURRENT,Y_SENSE_RESISTOR);
-#endif
-#ifdef Y2_IS_TMC
-	TMC26XStepper stepperY2(200,Y2_ENABLE_PIN,Y2_STEP_PIN,Y2_DIR_PIN,Y2_MAX_CURRENT,Y2_SENSE_RESISTOR);
-#endif
-#ifdef Z_IS_TMC
-	TMC26XStepper stepperZ(200,Z_ENABLE_PIN,Z_STEP_PIN,Z_DIR_PIN,Z_MAX_CURRENT,Z_SENSE_RESISTOR);
-#endif
-#ifdef Z2_IS_TMC
-	TMC26XStepper stepperZ2(200,Z2_ENABLE_PIN,Z2_STEP_PIN,Z2_DIR_PIN,Z2_MAX_CURRENT,Z2_SENSE_RESISTOR);
-#endif
-#ifdef E0_IS_TMC
-	TMC26XStepper stepperE0(200,E0_ENABLE_PIN,E0_STEP_PIN,E0_DIR_PIN,E0_MAX_CURRENT,E0_SENSE_RESISTOR);
-#endif
-#ifdef E1_IS_TMC
-	TMC26XStepper stepperE1(200,E1_ENABLE_PIN,E1_STEP_PIN,E1_DIR_PIN,E1_MAX_CURRENT,E1_SENSE_RESISTOR);
-#endif
-#ifdef E2_IS_TMC
-	TMC26XStepper stepperE2(200,E2_ENABLE_PIN,E2_STEP_PIN,E2_DIR_PIN,E2_MAX_CURRENT,E2_SENSE_RESISTOR);
-#endif
-#ifdef E3_IS_TMC
-	TMC26XStepper stepperE3(200,E3_ENABLE_PIN,E3_STEP_PIN,E3_DIR_PIN,E3_MAX_CURRENT,E3_SENSE_RESISTOR);
-#endif	
+
 //===========================================================================
 //=============================functions         ============================
 //===========================================================================
@@ -1420,49 +1390,3 @@ void microstep_readings()
       SERIAL_PROTOCOLLN( digitalRead(E1_MS2_PIN));
       #endif
 }
-
-#ifdef HAVE_TMCDRIVER
-void tmc_init()
-{
-#ifdef X_IS_TMC
-	stepperX.setMicrosteps(X_MICROSTEPS);
-	stepperX.start();
-#endif
-#ifdef X2_IS_TMC
-	stepperX2.setMicrosteps(X2_MICROSTEPS);
-	stepperX2.start();
-#endif
-#ifdef Y_IS_TMC
-	stepperY.setMicrosteps(Y_MICROSTEPS);
-	stepperY.start();
-#endif
-#ifdef Y2_IS_TMC
-	stepperY2.setMicrosteps(Y2_MICROSTEPS);
-	stepperY2.start();
-#endif
-#ifdef Z_IS_TMC
-	stepperZ.setMicrosteps(Z_MICROSTEPS);
-	stepperZ.start();
-#endif
-#ifdef Z2_IS_TMC
-	stepperZ2.setMicrosteps(Z2_MICROSTEPS);
-	stepperZ2.start();
-#endif
-#ifdef E0_IS_TMC
-	stepperE0.setMicrosteps(E0_MICROSTEPS);
-	stepperE0.start();
-#endif
-#ifdef E1_IS_TMC
-	stepperE1.setMicrosteps(E1_MICROSTEPS);
-	stepperE1.start();
-#endif
-#ifdef E2_IS_TMC
-	stepperE2.setMicrosteps(E2_MICROSTEPS);
-	stepperE2.start();
-#endif
-#ifdef E3_IS_TMC
-	stepperE3.setMicrosteps(E3_MICROSTEPS);
-	stepperE3.start();
-#endif
-}
-#endif
